@rez1dent3 2022-08-20
Действия по оптимизации
1. Regexp нужно компилировать 1 раз;
2. json unmarshal достаточно медленный;
3. Памяти на тесте TestGetDomainStat_Time_And_Memory не хватает, завернул пользователей в канал;
---

Профилировал на файле hw10_program_optimization/stats_test.go

Regexp.

CPU.
<img width="1512" alt="Снимок экрана 2022-08-20 в 20 08 44" src="https://user-images.githubusercontent.com/5111255/185758665-d848a5e1-d18c-4565-9fac-317a10459304.png">
<img width="1512" alt="Снимок экрана 2022-08-20 в 20 09 22" src="https://user-images.githubusercontent.com/5111255/185758673-fff772e0-f884-475a-aa15-b17885111177.png">
<img width="1512" alt="Снимок экрана 2022-08-20 в 20 09 32" src="https://user-images.githubusercontent.com/5111255/185758676-df90d24c-aa2e-4bce-9cfb-b78296e2bfff.png">

Memory.
<img width="1512" alt="Снимок экрана 2022-08-20 в 20 13 25" src="https://user-images.githubusercontent.com/5111255/185758821-b0961b4c-b3d5-4adc-9b24-825a3ed26bc2.png">
<img width="1512" alt="Снимок экрана 2022-08-20 в 20 13 52" src="https://user-images.githubusercontent.com/5111255/185758823-e91fce3c-edaf-4588-8d7a-b31dec41a84f.png">
<img width="1512" alt="Снимок экрана 2022-08-20 в 20 14 06" src="https://user-images.githubusercontent.com/5111255/185758825-08f2c2de-d2a1-4026-887b-f6fcbf2de462.png">

--
после повторял операции пока не прошли все тесты

<img width="1512" alt="Снимок экрана 2022-08-20 в 20 18 16" src="https://user-images.githubusercontent.com/5111255/185758991-fc9ce980-eb66-4e73-ac68-3039c84139cb.png">
<img width="1512" alt="Снимок экрана 2022-08-20 в 20 18 39" src="https://user-images.githubusercontent.com/5111255/185758995-42786df6-5df1-4f21-b1a1-51c1d50ecc1b.png">
<img width="1512" alt="Снимок экрана 2022-08-20 в 20 18 51" src="https://user-images.githubusercontent.com/5111255/185758996-27b26a59-1125-4973-9191-0cb3097826b4.png">